Move `consumers` directory to `subscription`
