+++ apps/subscription/src/modules/subscription/subscription.module.ts
@@ -9,7 +9,7 @@ import { PrismaModule } from '../../database/prisma.module';
 import { SubscriptionController } from './presentation/subscription.controller';
 import { WeatherClientModule } from '../clients/weather-client/weather-client.module';
 import { EmailProducerModule } from '@utils/modules/producers/email-producer/email-producer.module';
-import { SubscriptionConsumer } from '../consumers/subscription-consumer.service';
+import { SubscriptionConsumer } from './consumers/subscription-consumer.service';
 
 @Module({
   controllers: [SubscriptionController, SubscriptionConsumer],
